+++ tests/Keboola/Extractor/ApplicationTest.php
@@ -4,115 +4,91 @@ declare(strict_types=1);
 
 namespace Keboola\DbExtractor\Tests;
 
-use Keboola\DbExtractor\Test\ExtractorTest;
 use Keboola\Csv\CsvFile;
 use Symfony\Component\Process\Process;
 use Symfony\Component\Yaml\Yaml;
+use Keboola\DbExtractor\Exception\UserException;
 
-class ApplicationTest extends ExtractorTest
+class ApplicationTest extends BaseTest
 {
-    /** @var array */
-    private $dbConfig;
-
-    /** @var  string */
-    protected $rootPath;
-
-    /** @var string  */
-    protected $dataDir = __DIR__ . '/../../data';
-
-    public function setUp(): void
+    private function replaceConfig(array $config, string $format): void
     {
-        $this->rootPath = __DIR__ . '/../../..';
-
-        parent::setUp();
-        if (getenv('EXTERNAL_PG_HOST') === false) {
-            $this->fail("Missing environment var 'EXTERNAL_PG_HOST'");
+        @unlink($this->dataDir . '/config.json');
+        @unlink($this->dataDir . '/config.yml');
+        if ($format === self::CONFIG_FORMAT_JSON) {
+            file_put_contents($this->dataDir . '/config.json', json_encode($config));
+        } else if ($format === self::CONFIG_FORMAT_YAML) {
+            file_put_contents($this->dataDir . '/config.yml', Yaml::dump($config));
+        } else {
+            throw new UserException("Invalid config format type [{$format}]");
         }
-        if (getenv('EXTERNAL_PG_DATABASE') === false) {
-            $this->fail("Missing environment var 'EXTERNAL_PG_DATABASE'");
-        }
-        if (getenv('EXTERNAL_PG_USER') === false) {
-            $this->fail("Missing environment var 'EXTERNAL_PG_USER'");
-        }
-        if (getenv('EXTERNAL_PG_PASSWORD') === false) {
-            $this->fail("Missing environment var 'EXTERNAL_PG_PASSWORD'");
-        }
-        $this->dbConfig['host'] = getenv('EXTERNAL_PG_HOST');
-        $this->dbConfig['database'] = getenv('EXTERNAL_PG_DATABASE');
-        $this->dbConfig['user'] = getenv('EXTERNAL_PG_USER');
-        $this->dbConfig['password'] = getenv('EXTERNAL_PG_PASSWORD');
-        $this->dbConfig['port'] = (!is_null(getenv('EXTERNAL_PG_PORT'))) ? getenv('EXTERNAL_PG_PORT') : 5432;
     }
 
-    public function testRunAction(): void
+    /**
+     * @dataProvider configProvider
+     */
+    public function testRunAction(array $config, string $format): void
     {
-        $outputCsvFile = new CsvFile($this->dataDir . '/out/tables/in.c-main.info_schema.csv');
-        $manifestFile = $this->dataDir . '/out/tables/in.c-main.info_schema.csv.manifest';
-        @unlink($outputCsvFile->getPathname());
-        @unlink($manifestFile);
-
-        $config = Yaml::parse(file_get_contents($this->dataDir . '/pgsql/external_config.yml'));
-        $config['parameters']['db'] = $this->dbConfig;
-        @unlink($this->dataDir . '/config.yml');
-        file_put_contents($this->dataDir . '/config.yml', Yaml::dump($config));
-
+        $this->replaceConfig($config, $format);
         $process = Process::fromShellCommandline('php ' . $this->rootPath . '/run.php --data=' . $this->dataDir);
         $process->setTimeout(300);
-        $process->run();
+        $process->mustRun();
 
+        var_export($process->getErrorOutput());
+        var_export($process->getOutput());
         $this->assertEquals(0, $process->getExitCode());
         $this->assertEquals("", $process->getErrorOutput());
-
-        $this->assertTrue($outputCsvFile->isFile());
-        $this->assertFileExists($manifestFile);
     }
 
-    public function testTestConnectionAction(): void
+    /**
+     * @dataProvider configProvider
+     */
+    public function testTestConnectionAction(array $config, string $format): void
     {
-        $config = Yaml::parse(file_get_contents($this->dataDir . '/pgsql/external_config.yml'));
-        @unlink($this->dataDir . '/config.yml');
         $config['action'] = 'testConnection';
-        $config['parameters']['db'] = $this->dbConfig;
-        file_put_contents($this->dataDir . '/config.yml', Yaml::dump($config));
+        $this->replaceConfig($config, $format);
 
         $process = Process::fromShellCommandline('php ' . $this->rootPath . '/run.php --data=' . $this->dataDir);
         $process->setTimeout(300);
-        $process->run();
+        $process->mustRun();
         $this->assertJson($process->getOutput());
         $this->assertEquals(0, $process->getExitCode());
         $this->assertEquals("", $process->getErrorOutput());
     }
 
-    public function testTrailingSemicolonQuery(): void
+    /**
+     * @dataProvider configProvider
+     */
+    public function testTrailingSemicolonQuery(array $config, string $format): void
     {
-        $config = Yaml::parse(file_get_contents($this->dataDir . '/pgsql/external_config.yml'));
-        $config['parameters']['db'] = $this->dbConfig;
         $config['parameters']['tables'][0]['query'] = $config['parameters']['tables'][0]['query'] . ";";
-        @unlink($this->dataDir . '/config.yml');
-        file_put_contents($this->dataDir . '/config.yml', Yaml::dump($config));
-
+        $this->replaceConfig($config, $format);
         $process = Process::fromShellCommandline('php ' . $this->rootPath . '/run.php --data=' . $this->dataDir);
         $process->setTimeout(300);
-        $process->run();
+        $process->mustRun();
         $this->assertEquals(0, $process->getExitCode());
         $this->assertEquals("", $process->getErrorOutput());
     }
 
-    public function testProcessTimeout(): void
+    /**
+     * @dataProvider configProvider
+     */
+    public function testProcessTimeout(array $config, string $format): void
     {
-        $config = Yaml::parse(file_get_contents($this->dataDir . '/pgsql/external_config.yml'));
-        $config['parameters']['db'] = $this->dbConfig;
         $config['parameters']['tables'][0]['query'] = "SELECT pg_sleep(65), 1";
-        @unlink($this->dataDir . '/config.yml');
+        $this->replaceConfig($config, $format);
         file_put_contents($this->dataDir . '/config.yml', Yaml::dump($config));
 
         $process = Process::fromShellCommandline('php ' . $this->rootPath . '/run.php --data=' . $this->dataDir);
         $process->setTimeout(300);
-        $process->run();
+        $process->mustRun();
         $this->assertEquals(0, $process->getExitCode());
         $this->assertEquals("", $process->getErrorOutput());
     }
 
+    /**
+     * @dataProvider configProvider
+     */
     public function testUserError(): void
     {
         $config = Yaml::parse(file_get_contents($this->dataDir . '/pgsql/external_config.yml'));
@@ -137,8 +113,8 @@ class ApplicationTest extends ExtractorTest
         @unlink($outputCsvFile->getPathname());
         @unlink($manifestFile);
 
-        $config = Yaml::parse(file_get_contents($this->dataDir . '/pgsql/external_config.yml'));
-        $config['parameters']['db'] = $this->dbConfig;
+        $config = $this->getConfig();
+
         // queries with comments will break the () in the \copy command.
         // Failed \copy commands should fallback to using the old PDO method
         $config['parameters']['tables'][0]['query'] = "
@@ -147,8 +123,7 @@ class ApplicationTest extends ExtractorTest
                 select * from information_schema.columns
             ) as columns ON tables.table_schema = columns.table_schema AND tables.table_name = columns.table_name;
         ";
-        @unlink($this->dataDir . '/config.yml');
-        file_put_contents($this->dataDir . '/config.yml', Yaml::dump($config));
+        $this->replaceConfig($config, self::CONFIG_FORMAT_JSON);
 
         $process = Process::fromShellCommandline('php ' . $this->rootPath . '/run.php --data=' . $this->dataDir);
         $process->setTimeout(300);
@@ -162,17 +137,17 @@ class ApplicationTest extends ExtractorTest
         $this->assertFileExists($manifestFile);
     }
 
-    public function testGetTablesAction(): void
+    /**
+     * @dataProvider configProvider
+     */
+    public function testGetTablesAction(array $config, string $format): void
     {
-        $config = Yaml::parse(file_get_contents($this->dataDir . '/pgsql/external_config.yml'));
-        @unlink($this->dataDir . '/config.yml');
         $config['action'] = 'getTables';
-        $config['parameters']['db'] = $this->dbConfig;
-        file_put_contents($this->dataDir . '/config.yml', Yaml::dump($config));
+        $this->replaceConfig($config, $format);
 
         $process = Process::fromShellCommandline('php ' . $this->rootPath . '/run.php --data=' . $this->dataDir);
         $process->setTimeout(300);
-        $process->run();
+        $process->mustRun();
         $this->assertJson($process->getOutput());
 
         $this->assertEquals(0, $process->getExitCode());
